Make sure mask is float (to avoid wierd behavior.)

+++ captioning/modules/losses.py
@@ -174,7 +174,7 @@ class LanguageModelCriterion(nn.Module):
             mask = mask.reshape(-1, mask.shape[2])
         # truncate to the same size
         target = target[:, :input.size(1)]
-        mask =  mask[:, :input.size(1)]
+        mask =  mask[:, :input.size(1)].float()
 
         output = -input.gather(2, target.unsqueeze(2)).squeeze(2) * mask
         # Average over each token
@@ -203,7 +203,7 @@ class LabelSmoothing(nn.Module):
 
         input = input.reshape(-1, input.size(-1))
         target = target.reshape(-1)
-        mask = mask.reshape(-1)
+        mask = mask.reshape(-1).float()
 
         # assert x.size(1) == self.size
         self.size = input.size(1)
